reorganise a bit
