+++ src/Data/XML/Parse/Ordered.hs
@@ -16,7 +16,6 @@ import Control.Monad.State.Strict
 import Data.List (intercalate)
 import qualified Data.Map.Strict as Map
 import Data.Tuple
-import Data.XML.Parse.Class
 import Data.XML.Parse.Types
 import Data.XML.Types
 import Text.XML (Name)
